envtest/requirements.txt
alembic==1.11.1
black==23.3.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mpmath==1.3.0
mypy==1.4.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyngrok==6.0.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.18
sympy==1.12
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
urllib3==2.0.4
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0

